CAPI: skip unsuccefully syncked pictures KP-4188
Former-commit-id: b4a1c7f0d8f28e605402384e9e3002b50ec3fd6d

+++ src/UI/Capi/WB.UI.Capi/Syncronization/SynchronozationProcessor.cs
@@ -162,10 +162,11 @@ namespace WB.UI.Capi.Syncronization
                 {
                     this.ExitIfCanceled();
 
-                    this.push.PushBinary(this.credentials.Login, this.credentials.Password, binaryData.Data, binaryData.FileName,
+                    var result = this.push.PushBinary(this.credentials.Login, this.credentials.Password, binaryData.Data, binaryData.FileName,
                         binaryData.InterviewId, this.ct);
-
-                    this.fileSyncRepository.RemoveBinaryDataFromSyncFolder(binaryData.InterviewId, binaryData.FileName);
+                    
+                    if(result)
+                        this.fileSyncRepository.RemoveBinaryDataFromSyncFolder(binaryData.InterviewId, binaryData.FileName);
 
                     this.OnStatusChanged(new SynchronizationEventArgsWithPercent("pushing binary data", Operation.Push, true, (i * 100) / binaryDatas.Count));
                     i++;
